Add files via upload
This script creates a matrix that when you input the number of rows and columns it creates a matrix where the value of the first element of each row is the number of column and the value of the first element of each column is the number of the row. Every other element has the value of the element to the left added to the one above it.